security, html

+++ src/main/java/com/HPMS/HPMS/nurse/nursedto/NurseDTOController.java
@@ -6,9 +6,12 @@ import com.HPMS.HPMS.nurse.nurseinformation.NurseInformationService;
 import com.HPMS.HPMS.nurse.nursemain.NurseMain;
 import com.HPMS.HPMS.nurse.nursemain.NurseMainService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.security.Principal;
 import java.util.List;
@@ -23,7 +26,7 @@ public class NurseDTOController {
     private final NurseInformationService nurseInformationService;
     private final LicenseService licenseService;
 
-//    @PreAuthorize("hasAnyAuthority('ROLE_ADMIN','ROLE_SYSTEM','ROLE_DOCTOR','ROLE_NURSE')")
+    @PreAuthorize("hasAnyAuthority('ROLE_ADMIN','ROLE_SYSTEM','ROLE_DOCTOR','ROLE_NURSE')")
     @GetMapping("/info/{nurseId}")
     public String showNurse(@PathVariable Integer nurseId, Model model) {
         NurseMain nurseMain = nurseMainService.getNurseMain(nurseId);
@@ -45,7 +48,7 @@ public class NurseDTOController {
         return "nurse/nurse_information"; // 또는 "nurse/detail" 원하는 뷰로
     }
 
-//    @PreAuthorize("hasAnyAuthority('ROLE_ADMIN','ROLE_SYSTEM','ROLE_DOCTOR','ROLE_NURSE')")
+    @PreAuthorize("hasAnyAuthority('ROLE_ADMIN','ROLE_SYSTEM','ROLE_DOCTOR','ROLE_NURSE')")
     @GetMapping("/modify/{id}")
     public String showModifyForm(@PathVariable Integer id, Model model) {
         NurseDTO nurseDTO = nurseDTOService.getNurseDTO(id);
@@ -56,7 +59,7 @@ public class NurseDTOController {
         return "nurse/nurse_modify_form";
     }
 
-//    @PreAuthorize("hasAnyAuthority('ROLE_ADMIN','ROLE_SYSTEM','ROLE_DOCTOR','ROLE_NURSE')")
+    @PreAuthorize("hasAnyAuthority('ROLE_ADMIN','ROLE_SYSTEM','ROLE_DOCTOR','ROLE_NURSE')")
     @PostMapping("/modify/{id}")
     public String modifyNurse(@PathVariable Integer id, @ModelAttribute NurseDTO nurseDTO, Model model) {
         String modifier = nurseDTO.getNurseMain().getModifier();
@@ -69,13 +72,13 @@ public class NurseDTOController {
         return "redirect:/nurse/info/" + id;
     }
 
-//    @PreAuthorize("hasAnyAuthority('ROLE_ADMIN','ROLE_SYSTEM','ROLE_DOCTOR','ROLE_NURSE')")
+    @PreAuthorize("hasAnyAuthority('ROLE_ADMIN','ROLE_SYSTEM','ROLE_DOCTOR','ROLE_NURSE')")
     @GetMapping("/delete/{id}")
     public String nurseDelete(Principal principal, @PathVariable("id") Integer id) {
         NurseDTO nurseDTO = this.nurseDTOService.getNurseDTO(id);
-//        if (!nurseDTO.getNurseMain().getFirstName().equals(principal.getName())) {
-//            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "삭제권한이 없습니다.");
-//        }
+        if (!nurseDTO.getNurseMain().getFirstName().equals(principal.getName())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "삭제권한이 없습니다.");
+        }
         this.nurseDTOService.delete(nurseDTO);
         return "redirect:/nurse";
     }
